+++ src/audio_metadata/__about__.py
@@ -1,8 +1,12 @@
 __all__ = [
-	'__author__', '__author_email__',
-	'__copyright__', '__license__',
-	'__summary__', '__title__',
-	'__url__', '__version__',
+	'__author__',
+	'__author_email__',
+	'__copyright__',
+	'__license__',
+	'__summary__',
+	'__title__',
+	'__url__',
+	'__version__',
 	'__version_info__'
 ]
 
@@ -10,7 +14,7 @@ __title__ = 'audio-metadata'
 __summary__ = 'A library for reading and, in the future, writing metadata from audio files.'
 __url__ = 'https://github.com/thebigmunch/audio-metadata'
 
-__version__ = '0.4.0'
+__version__ = '0.5.0'
 __version_info__ = tuple(int(i) for i in __version__.split('.') if i.isdigit())
 
 __author__ = 'thebigmunch'